+++ sql-dsl/src/main/java/com/neaterbits/query/sql/dsl/api/IAdhoc.java
@@ -9,11 +9,11 @@ interface IAdhoc<MODEL>
 	IAdhocNumericTableResult<MODEL, Long>,
 	IAdhocNumericTableResult<MODEL, BigDecimal>> */ {
 	
-    <T> IAdhocNumericTableResult<MODEL, Short, T> sum(IFunctionShort<T> field);
+    <T> IAdhocNumericTableResult<MODEL, Integer, T> sum(IFunctionShort<T> field);    // Sum short to integer    (TODO: throw exception on overflow)
 	
-    <T> IAdhocNumericTableResult<MODEL, Integer, T> sum(IFunctionInteger<T> field);
-
-    <T> IAdhocNumericTableResult<MODEL, Long, T> sum(IFunctionLong<T> field);
+    <T> IAdhocNumericTableResult<MODEL, Integer, T> sum(IFunctionInteger<T> field);  // Sum integers to integer (TODO: throw exception on overflow) 
+    
+    <T> IAdhocNumericTableResult<MODEL, Long, T> sum(IFunctionLong<T> field);   	 // Sum integers to integer (TODO: throw exception on overflow)			
 
     <T> IAdhocNumericTableResult<MODEL, BigDecimal, T> sum(IFunctionBigDecimal<T> field);
